+++ docs/src/CardStackSection.tsx
@@ -1,0 +1,4 @@
+import {docs} from "../../packages/ods-lab/src/CardStackSection/CardStackSection.docs";
+import {renderDocs} from "./_private/renderDocs";
+
+export const CardStackSectionDocument = () => renderDocs(docs);