+++ docs/src/Section.tsx
@@ -1,0 +1,4 @@
+import {docs} from "../../packages/ods-core/src/Section/Section.docs";
+import {renderDocs} from "./_private/renderDocs";
+
+export const SectionDocument = () => renderDocs(docs);