+++ docs/src/routes/Documentation/foundations/index.ts
@@ -1,8 +1,0 @@
-import layout from "./layout/layout";
-import tones from "./tones/tones";
-
-// eslint-disable-next-line import/no-default-export
-export default {
-  "/foundations/layout": layout,
-  "/foundations/tones": tones,
-};