+++ docs/src/routes/Documentation/guides/index.ts
@@ -1,8 +1,0 @@
-import designWorkflow from "./design-workflow";
-import developmentWorkflow from "./development-workflow";
-
-// eslint-disable-next-line import/no-default-export
-export default {
-  "/guides/design-workflow": designWorkflow,
-  "/guides/development-workflow": developmentWorkflow,
-};