+++ packages/ods-core/src/Placeholder/Placeholder.tsx
@@ -1,7 +1,9 @@
 import React, {ReactNode} from "react";
 import styled from "styled-components";
+import {ColorVariants} from "@origin-digital/ods-themes";
 import {height, width} from "styled-system";
 import {Box, BoxProps} from "../Box/Box";
+import {Text} from "../Text/Text";
 
 const StyledBox = styled(Box)`
   ${height}
@@ -12,6 +14,7 @@ export interface PlaceholderProps {
   children?: ReactNode;
   "data-id"?: string;
   backgroundColor?: BoxProps["backgroundColor"];
+  color?: ColorVariants;
   inline?: boolean;
   height?: string;
   width?: string;
@@ -20,6 +23,7 @@ export interface PlaceholderProps {
 export const Placeholder = ({
   children,
   backgroundColor,
+  color,
   "data-id": dataId,
   inline,
   height,
@@ -37,16 +41,9 @@ export const Placeholder = ({
       height={height as BoxProps["height"]}
       width={width as BoxProps["height"]}
     >
-      <span
-        style={{
-          fontSize: "16px",
-          lineHeight: "24px",
-          fontWeight: 500,
-          color: "rgb(35, 35, 35)",
-        }}
-      >
+      <Text size="xsmall" color={color}>
         {children}
-      </span>
+      </Text>
     </StyledBox>
   );
 };