+++ playroom/CHANGELOG.md
@@ -3,12 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [0.4.0](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.3.7...playroom@0.4.0) (2020-02-16)
+## [0.3.7](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.3.6...playroom@0.3.7) (2020-02-12)
+
+**Note:** Version bump only for package playroom
 
 
-### Features
 
-* dynamically generate docs platform ([20659c9](https://bitbucket.orgn.io/od/origin-ui/commits/20659c9))
+
+
+## [0.3.6](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.3.5...playroom@0.3.6) (2020-02-12)
+
+**Note:** Version bump only for package playroom
 
 
 
@@ -26,21 +31,39 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package playroom
 
+
+
+
+
 ## [0.3.3](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.3.2...playroom@0.3.3) (2020-02-10)
 
 **Note:** Version bump only for package playroom
+
+
+
+
 
 ## [0.3.2](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.3.1...playroom@0.3.2) (2020-02-06)
 
 **Note:** Version bump only for package playroom
 
+
+
+
+
 ## [0.3.1](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.3.0...playroom@0.3.1) (2020-02-06)
 
 **Note:** Version bump only for package playroom
 
+
+
+
+
 # [0.3.0](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.2.5...playroom@0.3.0) (2020-02-05)
 
+
 ## [0.2.4](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.2.3...playroom@0.2.4) (2020-01-31)
+
 
 ## [0.2.3](https://bitbucket.orgn.io/od/origin-ui/compare/playroom@0.2.2...playroom@0.2.3) (2020-01-31)